+++ tests/cpp-tests/Classes/NetworkTest/HttpClientTest/HttpClientTest.cpp
@@ -311,11 +311,11 @@ void HttpClientTest::onHttpRequestCompleted(HttpClient *sender, HttpResponse *re
         log("%s completed", response->getHttpRequest()->getTag());
     }
     
-    long statusCode = response->getResponseCode();
+    int32_t statusCode = response->getResponseCode();
     char statusString[64] = {};
-    sprintf(statusString, "HTTP Status Code: %ld, tag = %s", statusCode, response->getHttpRequest()->getTag());
+    sprintf(statusString, "HTTP Status Code: %d, tag = %s", statusCode, response->getHttpRequest()->getTag());
     _labelStatusCode->setString(statusString);
-    log("response code: %ld", statusCode);
+    log("response code: %d", statusCode);
     
     if (response->getResponseCode() != 200) 
     {
@@ -460,11 +460,11 @@ void HttpClientClearRequestsTest::onHttpRequestCompleted(HttpClient *sender, Htt
         log("%s completed", response->getHttpRequest()->getTag());
     }
     
-    long statusCode = response->getResponseCode();
+    int32_t statusCode = response->getResponseCode();
     char statusString[64] = {};
-    sprintf(statusString, "HTTP Status Code: %ld, tag = %s", statusCode, response->getHttpRequest()->getTag());
+    sprintf(statusString, "HTTP Status Code: %d, tag = %s", statusCode, response->getHttpRequest()->getTag());
     _labelStatusCode->setString(statusString);
-    log("response code: %ld", statusCode);
+    log("response code: %d", statusCode);
     
     _totalProcessedRequests++;
     sprintf(statusString, "Got %d of %d expected http requests", _totalProcessedRequests, _totalExpectedRequests);
